# docs/swagger.yaml
basePath: /api/v1
definitions:
  comment_handler.HandleCreateCommentForTaskRequestParam:
    properties:
      task_id:
        type: integer
      text:
        type: string
    required:
    - task_id
    - text
    type: object
  dto.Comment:
    properties:
      author:
        $ref: '#/definitions/dto.User'
      created_at:
        type: string
      id:
        type: integer
      task_id:
        type: integer
      text:
        type: string
      updated_at:
        type: string
    type: object
  dto.Kanban:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      updated_at:
        type: string
    type: object
  dto.Notification:
    properties:
      created_at:
        description: omitempty, чтобы не возвращать null
        type: string
      description:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      owner_id:
        type: integer
      title:
        type: string
      updated_at:
        description: omitempty, чтобы не возвращать null
        type: string
    type: object
...
  dto.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      surname:
        type: string
      updated_at:
        type: string
    type: object
  kanban_handler.CreateKanbanRequestParam:
    properties:
      name:
        type: string
      project_id:
        type: integer
    required:
    - name
    - project_id
    type: object
  kanban_handler.UpdateKanbanRequestParam:
    properties:
      name:
        type: string
    type: object
  notification_handler.HandleCreateNotificationParam:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
...
  response.JSONResponse:
    description: default JSON-response
    properties:
      data:
        type: object
      error:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Task manager
  title: DataTask
  version: "1.0"
paths:
...
  /comment/forTask:
    post:
      consumes:
      - application/json
      description: Create a new comment for task
      parameters:
      - description: Comment data for creation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/comment_handler.HandleCreateCommentForTaskRequestParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Create Comment For Task
      tags:
      - Comment
  /comment/forTask/{task_id}:
    get:
      consumes:
      - application/json
      description: Get all comments for a specific task
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Comment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Comments By Task ID
      tags:
      - Comment
  /kanban:
    post:
      consumes:
      - application/json
      description: Create a new Kanban board
      parameters:
      - description: Kanban board data for creation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kanban_handler.CreateKanbanRequestParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Kanban'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Create Kanban board
      tags:
      - Kanban
  /kanban/{id}:
    delete:
      description: Delete a Kanban board by its ID
      parameters:
      - description: Kanban board ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Delete Kanban board
      tags:
      - Kanban
    get:
      description: Get a Kanban board by its ID
      parameters:
      - description: Kanban board ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Kanban'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Kanban board by ID
      tags:
      - Kanban
    put:
      consumes:
      - application/json
      description: Update a Kanban board's details
      parameters:
      - description: Kanban board ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Kanban board data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kanban_handler.UpdateKanbanRequestParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Kanban'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Update Kanban board
      tags:
      - Kanban
  /kanban/project/{id}:
    get:
      description: Get a Kanban board by project ID
      parameters:
      - description: Kanban board project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Kanban'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Kanban board by project ID
      tags:
      - Kanban
...
  /notification:
    get:
      description: Retrieves all notifications for a specific user
      produces:
      - application/json
      responses:
        "200":
          description: List of notifications
          schema:
            items:
              $ref: '#/definitions/dto.Notification'
            type: array
        "400":
          description: Invalid owner ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user notifications
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Creates a new notification for a user
      parameters:
      - description: Notification object to be created
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification_handler.HandleCreateNotificationParam'
      produces:
      - application/json
      responses:
        "201":
          description: Notification created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new notification
      tags:
      - Notifications
  /notification/{notification_id}/read:
    patch:
      description: Marks a specific notification as read by its ID
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notification marked as read successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid notification ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Notification not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mark notification as read
      tags:
      - Notifications
...
security:
- Authorization: []
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"